updated app.tsx to use public chat
